requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
mysql-connector==2.2.9
paramiko==2.11.0
pycparser==2.21
pydotenv==0.0.7
PyNaCl==1.5.0
python-dotenv==0.20.0
six==1.16.0
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
mysql-connector==2.2.9
paramiko==2.11.0
pycparser==2.21
pycryptodome==3.15.0
pydotenv==0.0.7
PyNaCl==1.5.0
python-dotenv==0.20.0
six==1.16.0
bcrypt==3.2.2
cffi==1.15.1
colored==1.4.3
commonmark==0.9.1
cryptography==37.0.4
mysql-connector==2.2.9
paramiko==2.11.0
pycparser==2.21
pycryptodome==3.15.0
pydotenv==0.0.7
Pygments==2.13.0
PyNaCl==1.5.0
python-dotenv==0.20.0
rich==12.5.1
six==1.16.0
